requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
